+++ templates/parent.h
@@ -36,7 +36,7 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
 namespace matchmaker
 {
-    int64_t size_snth();
-    std::string const & at_snth(int64_t const & index);
-    int64_t lookup_snth(std::string const & str, bool * found);
+    int size_snth();
+    std::string const & at_snth(int const & index);
+    int lookup_snth(std::string const & str, bool * found);
 }
